% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do.rbind.r
\name{do.rbind}
\alias{do.rbind}
\title{Bind data.frames in a List by Rows}
\usage{
do.rbind(x, idcol = "Name", keep.rownames = FALSE)
}
\arguments{
\item{x}{List with data.frames. Non data.frame objects are dropped.}

\item{idcol}{Name for column with ids in output.}

\item{keep.rownames}{Logical, keep rownames.}
}
\value{
Returns a data.frame
}
\description{
Same as \code{do.call(rbind, x)}, but adding a column with the
name of each table. Missing names are replaced by integers.
}
\author{
Sven E. Templer
}
