% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factorNA.r
\name{factorNA}
\alias{factorNA}
\title{Create a Factor with NA as Level}
\usage{
factorNA(x, ...)
}
\arguments{
\item{x}{A vector coerced to character.}

\item{...}{Forwarded to \link{factor}. \code{x} and \code{levels} are defined.}
}
\description{
Create a \link{factor} with \code{NA} values included
and positioned as last level.
}
\author{
Sven E. Templer
}
