% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help.index.R
\name{help.index}
\alias{help.index}
\title{Open The Package Help Index Page}
\usage{
help.index(pkg, browser = NULL)
}
\arguments{
\item{pkg}{A character string or expression with the name of a package.}

\item{browser}{The browser to display. \code{text} and \code{pdf} don't
use a browser, but builtin text/pdf (help_type). Otherwise a character
string for the browser program binary to call or function.}
}
\description{
Given a package name or string, start the package help index page
in a browser.
}
\author{
Sven E. Templer
}
\keyword{browser}
\keyword{help}
\keyword{index}
