% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/info.R
\name{info}
\alias{info}
\title{Print enhanced session information}
\usage{
info(..., width = 120)
}
\arguments{
\item{...}{Forwarded to other methods.}

\item{width}{Console width in columns.}
}
\description{
Based on and enhancing \code{devtools::session_info}.
}
\examples{
info()
devtools::session_info()
sessionInfo()
}
\seealso{
\link[devtools]{session_info}
}
\author{
Sven E. Templer
}
\keyword{session}
