% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsall.R
\name{lsall}
\alias{lsall}
\title{List Object Details}
\usage{
lsall(envir = .GlobalEnv, ...)
}
\arguments{
\item{envir}{An environment where to look for objects.}

\item{...}{Arguments forwarded to \code{ls}.}
}
\value{
Returns a \code{data.frame} with object names, lengths, classes, modes
and sizes or \code{NULL} if the environment is empty.
}
\description{
Return a data.frame with a list of all objects of a specified environmet.
}
\examples{
#

lsall()
obj1 <- 1:3
obj2 <- data.frame(1:3)
obj3 <- list(1:3)
lsall()

#
}
\seealso{
\link{ls}
}
\author{
Sven E. Templer
}
\keyword{all}
\keyword{list}
