\docType{package}
\name{miscset-package}
\alias{miscset}
\alias{miscset-package}
\title{Miscellaneous R Tools}
\description{
Set of miscellaneous R tools to simplify the work with
several data types and formats.
}
\details{
The package vignette provides a comprehensive overview and
examples for the usage of all available functions in the
package. Open with \code{vignette("miscset")}.
}
\author{
Sven E. Templer \email{sven.templer@gmail.com}
}
\keyword{misc}
\keyword{miscellaneous}
\keyword{tools}

