% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ntri}
\alias{ntri}
\title{Return Triangular Numbers}
\usage{
ntri(n)
}
\arguments{
\item{n}{Positive integer value for sequence length.}
}
\value{
Returns an integer vector of length \code{n}.
}
\description{
Return the series of triangular (/triangle) numbers up to a number of
\code{n} rows of a triangle. The series has the entry number "A000217"
at \url{https://oeis.org/A000217} and starts like this:
0, 1, 3, 6, 10, ...
}
\author{
Sven E. Templer (\email{sven.templer@gmail.com})
}
\keyword{number}
\keyword{triangle}
\keyword{triangular}
