% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nunique.r
\name{nunique}
\alias{nunique}
\alias{uniquei}
\title{Amount and Index of Unique Values}
\usage{
nunique(x, na = TRUE, ...)

uniquei(x, na = TRUE, ...)
}
\arguments{
\item{x}{Numeric vector to transform.}

\item{na}{Logical, \code{TRUE} to include/count \code{NA}.}

\item{...}{Arguments forwarded to \link{unique}.}
}
\description{
Return the index or amount of unique values in a vector.
}
\examples{
#

v <- c("a","b","a", NA)
nunique(v)
nunique(v, FALSE)
uniquei(v)
uniquei(v, FALSE)

#
}
\author{
Sven E. Templer
}
\keyword{unique}
