% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p2star.R
\name{p2star}
\alias{p2star}
\title{P Value Significance Level Indicator}
\usage{
p2star(p, breaks = c(0, 0.001, 0.01, 0.05, 0.1, 1), symbols = c("***", "**",
  "*", ".", "n.s."))
}
\arguments{
\item{p}{Vector with p values}

\item{breaks}{The breaks from min (0) to max (1).}

\item{symbols}{Symbols to use for values between breaks from min to max.}
}
\description{
Transform p-values to character (e.g. stars) indicators
by significance levels with the function \link{symnum}.
}
\examples{
#

p2star(c(1e-5,.1,.9))

#
}
\author{
Sven E. Templer
}
