% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotn.R
\name{plotn}
\alias{plotn}
\title{Plot Nothing (but a Plot)}
\usage{
plotn(x = 0:1, y = NULL, type = "n", xlab = "", ylab = "",
  xaxt = "n", yaxt = "n", frame.plot = F, ...)
}
\arguments{
\item{x}{Coordinates of the points.}

\item{y}{Coordinates of the y-axis.}

\item{type}{Plot type.}

\item{xlab, ylab}{Axis titles.}

\item{xaxt, yaxt}{Axis types.}

\item{frame.plot}{Plot the frame.}

\item{...}{Forwarded arguments to \code{plot}.}
}
\description{
Create a plot, with empty elements by presetting default parameters.
}
\details{
For details about the function see \link{plot}, which is called from
\code{plotn}. More detailed information in \link{plot.default} and
\link{par}.
}
\author{
Sven E. Templer
}
