% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmall.R
\name{rmall}
\alias{rmall}
\title{Remove All Objects from Global Environment}
\usage{
rmall(...)
}
\arguments{
\item{...}{Arguments forwarded to \code{ls} to get all objects.}
}
\description{
Remove all objects from the global environment.
}
\examples{
#

a <- b <- letters
ls()
rmall()
ls()

#
}
\seealso{
\link{rm}, \link{ls}
}
\author{
Sven E. Templer
}
\keyword{all}
\keyword{remove}
