% Generated by roxygen2 (4.0.2): do not edit by hand
\name{scale0}
\alias{scale0}
\title{Scale to Range 0 to 1}
\usage{
scale0(x)
}
\arguments{
\item{x}{Numeric vector to transform.}
}
\description{
Scale numeric values to a range from 0 to 1.
}
\examples{
#

scale0(0:10)
scale0(-1:3)
scale0(2:3)

#
}
\author{
Sven E. Templer \email{sven.templer@gmail.com}
}
\keyword{scale}

