% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/squarematrix.R
\name{squarematrix}
\alias{squarematrix}
\title{Create a Square Matrix}
\usage{
squarematrix(x)
}
\arguments{
\item{x}{Object of class \code{matrix}.}
}
\value{
Returns a \code{matrix}.
}
\description{
Transform any m x n matrix to a square matrix by column/row names.
Stops if no or duplicated dimnames are provided in x.
}
\examples{
#

m <- matrix(1:6, 2, dimnames=list(2:3,1:3))
m
squarematrix(m)

#
}
\author{
Sven E. Templer
}
\keyword{matrix}
\keyword{square}
