% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strrev.R
\name{str_rev}
\alias{str_rev}
\alias{strrev}
\title{Reverse Text Strings}
\usage{
str_rev(x)

strrev(x)
}
\arguments{
\item{x}{vector with strings. Is coerced to character.}
}
\value{
Returns a character vector with reversed strings.
}
\description{
Create a reverse version of strings.
}
\examples{
#

s <- c("abc","asdf")

str_rev(s)

#
}
\seealso{
\link{rev}
}
\author{
Sven E. Templer
}
\keyword{inverse}
\keyword{reverse}
