% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strrev.R
\name{strrev}
\alias{strrev}
\title{Reverse Text Strings}
\usage{
strrev(x)
}
\arguments{
\item{x}{vector with strings. Is coerced to character.}
}
\value{
Returns a character vector with reversed strings.
}
\description{
Create a reverse version of strings.
}
\examples{
#

strrev(c("abc","asdf"))

#
}
\author{
Sven E. Templer
}
\seealso{
\link{rev}
}
\keyword{inverse}
\keyword{reverse}

