% Generated by roxygen2 (4.0.2): do not edit by hand
\name{textable}
\alias{textable}
\title{Table to Latex}
\usage{
textable(d, file, as.document = TRUE, landscape = FALSE, caption = NULL,
  rownames = TRUE, align = NULL, pt.size = 10, margin = 2,
  digits = NULL, label = NULL, cmd = NULL)
}
\arguments{
\item{d}{Object to transform to a latex table.}

\item{file}{Character string with output file name. If missing or \code{""},
the output is printed to the screen.}

\item{as.document}{Logical. \code{TRUE} to add the document definition to
the output. The document class is an article and the package a4paper is
included.}

\item{landscape}{Logical, use a landscape format for wider tables.}

\item{caption}{Character vector with title of table.}

\item{rownames}{Logical, include row names of \code{d}.}

\item{align}{Character vector with \code{'l'}, \code{'c'}, \code{'r'}
for aligning the columns left, centered or right. Length is either one
or 1 (for rownames column) + number of columns in \code{d} (even if
\code{rownames = FALSE})}

\item{pt.size}{Integer from 10 to 13 for the size of the characters.}

\item{margin}{Margin between table and page border in cm.}

\item{digits}{Number of digits to print from numeric values.}

\item{label}{Character vector with the latex label or HTML anchor.}

\item{cmd}{A character vector with the system command to apply
on the output file. Only if \code{file} is given and \code{as.document}
is \code{TRUE}. \code{NULL} or an empty string does not call \link{system}.}
}
\value{
Returns a character vector invisible. If \code{file} is set, then the
content is written to a file. Else it is printed to the console.
}
\description{
This function enhances \code{xtable}: It can write the latex code of the
table directly to a file and optionally adds a header/footer for the
document structure. Also a system command can be given to convert the
tex file to a pdf document, for example.
}
\details{
Example for a system call:\cr
\code{cmd = "pdflatex -output-directory /path/to/files/"}
}
\examples{
#

textable(head(trees), rownames=TRUE, digits=4, caption='R "trees" dataset.')

#
}
\author{
Sven E. Templer (\email{sven.templer@gmail.com})
}
\seealso{
\link{xtable}
}

