% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ciplot.R
\name{ciplot}
\alias{ciplot}
\alias{ciplot.default}
\title{Barplot with confindence intervals}
\usage{
ciplot(x, ...)

\method{ciplot}{default}(x, height.fun = mean, range.fun = ci, ...,
  arrows.args = list())
}
\arguments{
\item{x}{List of numeric values}

\item{range.fun}{Function to calculate the range of the
arrows (e.g. a confidence interval). See details.}

\item{heigth.fun}{Function to apply on each list object
to calculate the height of the bars from.}
}
\description{
Create barplots of a list of numeric values and arrows
according to the confidence interval, standard deviation,
interquartile range, etc.
}
\details{
sd_range <- function (x) { m <- mean(x); s <- sd(x); c(m-s,m+s)}
}
\author{
Sven E. Templer
}

