% Generated by roxygen2 (4.0.2): do not edit by hand
\name{help.index}
\alias{help.index}
\title{Open The Package Help Index Page in a Browser}
\usage{
help.index(pkg, browser = NA, encodeIfNeeded = FALSE)
}
\arguments{
\item{pkg}{A character string or variable name of a package.}

\item{browser}{A character string for a browser command to
replace the option 'browser' with. Use the string "rstudio"
for calling the internal browser in RStudio.}

\item{encodeIfNeeded}{Logical value. \code{TRUE} to encode
the url. Mostly not necessary. Read More at \link{browseURL}.}
}
\description{
Given a package name or string, start the package help index page
in a browser.
}
\author{
Sven E. Templer (\email{sven.templer@gmail.com})
}
\keyword{browser}
\keyword{help}
\keyword{index}

