% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa.R
\docType{package}
\name{miscset-package}
\alias{miscset-package}
\alias{miscset}
\title{Miscellaneous R Tools}
\description{
A collection of miscellaneous methods to simplify various tasks,
including plotting, data.frame and matrix transformations, environment
functions, regular expression methods, and string and logical operations, as
well as numerical and statistical tools. Most of the methods are simple but
useful wrappers of common base R functions, which extend S3 generics or
provide default values for important parameters.
}
\details{
The package vignette provides a comprehensive overview and examples for the
usage of all available functions in the package. View with 
\code{vignette("miscset")}.
}
\author{
Sven E. Templer
}
\keyword{misc}
\keyword{miscellaneous}
\keyword{tools}
