% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/aaa.R
\docType{package}
\name{miscset-package}
\alias{miscset}
\alias{miscset-package}
\title{Miscellaneous R Tools}
\description{
Miscellaneous functions to simplify R workspace and data handling. This includes numeric value scaling, string manipulation by regular expressions, plotting 'ggplot' arrays or empty plots, reshaping matrices and applying functions by a grid on data.frames. Furthermore it contains functions to maintain workspaces by enhanced object listing or removing and multiple RData file loading. For performance, multi-core support and C++ functions are implemented.
}
\details{
The package vignette provides a comprehensive overview and examples for the usage
of all available functions in the package. Open with \code{vignette("miscset")}.
}
\author{
Sven E. Templer \email{sven.templer@gmail.com}
}
\keyword{misc}
\keyword{miscellaneous}
\keyword{tools}

