% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa.R
\docType{package}
\name{miscset-package}
\alias{miscset}
\alias{miscset-package}
\title{Miscellaneous R Tools}
\description{
Miscellaneous methods that simplify tasks from multiple areas. 
This comprises plotting, data.frame and matrix transformations, environment functions, 
regular expression methods, string and logical operations and numerical 
and statistical tools. Most of the methods are simple but useful wrappers 
of common base R functions by extending S3 generics or setting default parameters.
}
\details{
The package vignette provides a comprehensive overview and examples for the
usage of all available functions in the package. View with 
\code{vignette("miscset")}.
}
\author{
Sven E. Templer
}
\keyword{misc}
\keyword{miscellaneous}
\keyword{tools}

