% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/aaa.R
\docType{package}
\name{miscset-package}
\alias{miscset}
\alias{miscset-package}
\title{Miscellaneous R Tools}
\description{
The package contains methods to simplify workspace handling, sort,
reshape and apply functions on grids of data.frames, scale numeric values,
extract unique and duplicate values, perform regular expression based
string operations and ease plotting. Many methods are implemented with
multi-core support from the parallel package or written in C++.
}
\details{
The package vignette provides a comprehensive overview and examples for the
usage of all available functions in the package. View with
\code{vignette("miscset")}.
}
\author{
Sven E. Templer \email{sven.templer@gmail.com}
}
\keyword{misc}
\keyword{miscellaneous}
\keyword{tools}

