\name{rmall}
\alias{rmall}
\title{Remove All Objects from Global Environment}
\usage{
rmall(...)
}
\arguments{
  \item{...}{Arguments forwarded to \code{ls} to get all
  objects.}
}
\description{
Remove all objects from the global environment.
}
\author{
Sven E. Templer \email{sven.templer@gmail.com}
}
\seealso{
\link{rm}, \link{ls}
}
\keyword{all}
\keyword{remove}

