% Generated by roxygen2 (4.0.1): do not edit by hand
\name{sortable}
\alias{sortable}
\title{Return Sorted Table}
\arguments{
\item{x}{Vector to create table of counts}

\item{decr}{Decreased order if \code{TRUE}.}

\item{na}{Include \code{NA} if \code{TRUE}.}
}
\value{
Named integer Vector \code{n}.
}
\description{
R/C++ version of \code{sort(table(x))}, with the option to switch between
increasing or decreasing order and to include counts of \code{NA}.
}
\author{
Sven E. Templer (\email{sven.templer@gmail.com})
}
\keyword{sort}
\keyword{table}

